requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
certifi==2025.11.12
cffi==2.0.0
cfgv==3.5.0
click==8.3.1
cryptography==46.0.3
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.123.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.2
fastar==0.8.0
filelock==3.20.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
identify==2.6.15
idna==3.11
jinja2==3.1.6
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
nodeenv==1.9.1
passlib==1.7.4
pip==25.2
platformdirs==4.5.0
pre-commit==4.5.0
psycopg==3.3.0
psycopg-binary==3.3.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-core==2.41.5
pygments==2.19.2
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pyyaml==6.0.3
rich==14.2.0
rich-toolkit==0.17.0
rignore==0.7.6
rsa==4.9.1
sentry-sdk==2.46.0
shellingham==1.5.4
six==1.17.0
sqlalchemy==2.0.44
starlette==0.50.0
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
# uvicorn==0.38.0
uvloop==0.22.1
virtualenv==20.35.4
watchfiles==1.1.1
websockets==15.0.1
